requirements.txt
flask==3.1.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
mysql-connector-python==8.4.0
gunicorn==22.0.0
